point string decode
